+++ Puko/Controllers/main.php
@@ -1,6 +1,9 @@
 <?php
 
-class main
+use Puko\Core\Service\Service;
+use Puko\Core\View\View;
+
+class Main extends View
 {
 
     var $id;
@@ -12,6 +15,7 @@ class main
 
     function main()
     {
+        var_dump($this->id);
         return array(
             'PageTitle' => 'Puko Framework',
             'Welcome' => 'uiie',