+++ Puko/Core/Data.php
@@ -18,7 +18,7 @@ class Data
 
     private function __construct($tablename = null)
     {
-        $conn = (require_once PUKO_ROOT . '/Config/db.php');
+        $conn = (require_once ROOT . '/Config/db.php');
         $this->pdo = new PDO("mysql:host=" . $conn['host'] . ";dbname=" . $conn['dbName'], $conn['user'],
             $conn['pass']);
         $this->pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);