+++ Puko/Core/View/View.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Puko\Core\View;
+
+class View
+{
+
+}